Move Append test to strings_test.go.
